Drop pyexiv2 in favor of GExiv2.
The reason for this is that it provides a migration path towards
Python 3.

Note that this branch can't be merged into master or released until
libgexiv2 upstream merges my patches to support introspection.

Also, altitudes seem to be broken. Not sure why, since the same
method is correctly giving me latitudes and longitudes.