Drop PKG-INFO file.
This file is produced by distutils 'sdist' command for producing
source tarballs, and marcoil had originally been using sdist as the
official source of "upstream tarballs" for packaging, however I
consider github's tagged version tarballs as the official "upstream
tarballs", and they don't contain that file at all, so PKG-INFO has
become an ugly patch on top of the master branch.

I'm in the process of unifying the master and debian branches as
much as possible, with the goal of ultimately having no differences
between them, so that the 'debian' directory can just be dropped in
and packages built without any modifications to any other files.