Drop build dependency on git when building debian packages.
This is necessary for launchpad daily builds, which build
from a bazaar mirror of the official git master, and wouldn't have
the git history available even if we specify that git is a build
dependency.

This leaves in place the build dependency on git if you're building
from git master directly.