Drop the Makefile because it was interfering with dh_auto_build.
I wasn't able to find any way for dh_auto_build to ignore the Makefile
and keep using setup.py like it's always been, so I just wrote a
quick&dirty little addition to setup.py that lets it help me run
my tests, which is all I wanted out of the Makefile anyway.

Not as elegant as a makefile unfortunately, but it sure eliminates
a lot of headaches with dh_auto_build.